Unzipping *.egg files residing in eggs directory
One of project's dependencies really wanted setuptools to be uploaded
to App Engine, but it's downloaded as a zipped .egg file. This code
extracts zipped eggs files next to themselves and uses extracted
version when copying libraries.

I first tried to replace zipped egg with a directory, but apparently
zipped version is required by buildout.